dermdetect/requirement.txt
numpy==1.24.0
flask==2.2.2
keras==2.15.0
flask_cors==4.0.0
pillow==10.0.1